parser: properly support 64-bit numbers
The language supports specifying 64-bit unsigned values, but the num
field of the token structure is not guaranteed to be 64 bits wide.
Change its type to be unsigned long long, and change the code that
parses numbers to use strtoull() so we can actually accept a 64-bit
value.

This is also true of the value field of the token type.  Change it
to unsigned long long as well (and format it as unsigned).

Check the return value of strtoull(), and if the result is out of
range, report an error.

Message-Id: <20211001232338.769309-23-elder@linaro.org>

+++ qmic.h
@@ -28,7 +28,7 @@ extern const char *qmi_package;
 
 struct qmi_const {
 	const char *name;
-	unsigned value;
+	unsigned long long value;
 
 	struct list_head node;
 };
